feat/adding some new pages
