+++ index.js
@@ -3,6 +3,9 @@ const mongoose = require('mongoose')
 const config = require('./config/config')[env];
 const express = require('express');
 const indexRouter = require('./routes');
+const cubeRouter = require('./routes/cube')
+const accessoriesRouter = require('./routes/accessories')
+const authRouter = require('./routes/auth')
 
 const app = express();
 
@@ -24,4 +27,14 @@ require('./config/express')(app);
 
 
 app.use('/', indexRouter);
+app.use('/', cubeRouter);
+app.use('/', accessoriesRouter);
+app.use('/', authRouter);
+
+app.get('*', (req, res) =>{
+    res.render('404', {
+        title: 'Oh nooo!'
+    })
+})
+
 app.listen(config.port, console.log(`Listening on port ${config.port}! Now its up to you...`));
