+++ views/components/header.hbs
@@ -2,8 +2,11 @@
     <ul>
         <li><img class="logo" src="/static/images/logo.png"></li>
         <li><a href="/">Browse</a></li>
+        <li><a href="/about">About</a></li>
         <li><a href="/create">Add a Cube</a></li>
         <li><a href="/create/accessory">Add Accessory</a></li>
-        <li><a href="/about">About</a></li>
+        <li><a href="/login">Login</a></li>
+        <li><a href="/register">Register</a></li>
+        <li><a href="/">Logout</a></li>
     </ul>
 </header>
