+++ views/deleteCubePage.hbs
@@ -1,0 +1,21 @@
+<h1>Delete Cube</h1>
+<div class="form">
+<form action='/' method="POST">
+  <label for="name">Name</label>
+  <input type="text" id="name" name="name" value="" disabled="true"/>
+  <label for="name">Description</label>
+  <textarea id="name" name="description" disabled="true"/></textarea>
+  <label for="image">ImageUrl</label>
+  <input type="text" id="image" name="imageUrl" value="" disabled="true"/>
+  <label for="difficulty">Difficulty</label>
+  <select id="difficulty" name="difficultyLevel" value="" disabled="true"/>
+    <option value="1">1 - Very Easy</option>
+    <option value="2">2 - Easy</option>
+    <option value="3">3 - Medium (Standard 3x3)</option>
+    <option value="4">4 - Intermediate</option>
+    <option value="5">5 - Expert</option>
+    <option value="6">6 - Hardcore</option>
+  </select>
+  <input type="submit" value="Delete">
+</form>
+</div>
